chore(mobile): Fixes GroupWelcome check redirect with URQL Me query updater. Lints some code
