+++ apps/mobile/src/screens/Stream/Stream.js
@@ -3,12 +3,13 @@ import { useNavigation, useIsFocused } from '@react-navigation/native'
 import { View, TouchableOpacity, Dimensions } from 'react-native'
 import { useTranslation } from 'react-i18next'
 import { FlashList } from '@shopify/flash-list'
-import { gql, useMutation, useQuery } from 'urql'
+import { useMutation, useQuery } from 'urql'
 import { capitalize, get, isEmpty } from 'lodash/fp'
 import { clsx } from 'clsx'
 import { MY_CONTEXT_SLUG, PUBLIC_CONTEXT_SLUG } from '@hylo/shared'
 import { isStaticContext } from '@hylo/presenters/GroupPresenter'
 import updateUserSettingsMutation from '@hylo/graphql/mutations/updateUserSettingsMutation'
+import updateMembershipMutation from '@hylo/graphql/mutations/updateMembershipMutation'
 import useCurrentUser from '@hylo/hooks/useCurrentUser'
 import useCurrentGroup from '@hylo/hooks/useCurrentGroup'
 import useStreamQueryVariables from '@hylo/hooks/useStreamQueryVariables'
@@ -56,23 +57,6 @@ export const EVENT_STREAM_TIMEFRAME_OPTIONS = [
 export const DEFAULT_SORT_BY_ID = 'updated'
 export const DEFAULT_TIMEFRAME_ID = 'future'
 
-// Currently unused
-export const resetGroupTopicNewPostCountMutation = gql`
-  mutation ResetGroupTopicNewPostCountMutation($id: ID) {
-    updateGroupTopicFollow(id: $id, data: { newPostCount: 0 }) {
-      success
-    }
-  }
-`
-
-export const resetGroupNewPostCountMutation = gql`
-  mutation ResetGroupNewPostCountMutation ($id: ID) {
-    updateMembership(groupId: $id, data: { newPostCount: 0 }) {
-      id
-    }
-  }
-`
-
 export default function Stream () {
   const ref = useRef(null)
   const { t } = useTranslation()
@@ -116,7 +100,7 @@ export default function Stream () {
   const postIds = posts?.map(p => p.id)
 
   const [, updateUserSettings] = useMutation(updateUserSettingsMutation)
-  const [, resetGroupNewPostCount] = useMutation(resetGroupNewPostCountMutation)
+  const [, resetGroupNewPostCount] = useMutation(updateMembershipMutation)
 
   const title = useMemo(() => {
     if (customView?.name) {
@@ -155,7 +139,7 @@ export default function Stream () {
         sortBy === DEFAULT_SORT_BY_ID &&
         !streamQueryVariables.filter
       ) {
-        resetGroupNewPostCount({ id: currentGroup?.id })
+        resetGroupNewPostCount({ groupId: currentGroup?.id, data: { newPostCount: 0} })
       }
     }
   }, [currentGroup?.id, streamQueryVariables?.filter, streamQueryVariables?.context, hasMore, isFocused, postIds])
